requirements.txt
fastapi
uvicorn
sqlalchemy
aiomysql  # asyncmy choose one
async-exit-stack
async-generator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv

